# App/BackEnd/Conduit.API/src/Conduit.API/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "name" : "RealWorld",
      "url" : "https://realworld.io"
    },
    "description" : "Conduit API",
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "title" : "Conduit API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/api"
  } ],
  "paths" : {
    "/users/login" : {
      "post" : {
        "description" : "Login for existing user",
        "operationId" : "Login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginUserRequest"
              }
            }
          },
          "description" : "Credentials to use",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "summary" : "Existing user login",
        "tags" : [ "User and Authentication" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/users" : {
      "post" : {
        "description" : "Register a new user",
        "operationId" : "CreateUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewUserRequest"
              }
            }
          },
          "description" : "Details of the new user to register",
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "summary" : "Register a new user",
        "tags" : [ "User and Authentication" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/user" : {
      "get" : {
        "description" : "Gets the currently logged-in user",
        "operationId" : "GetCurrentUser",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Get current user",
        "tags" : [ "User and Authentication" ]
      },
      "put" : {
        "description" : "Updated user information for current user",
        "operationId" : "UpdateCurrentUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "description" : "User details to update. At least **one** field is required.",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Update current user",
        "tags" : [ "User and Authentication" ],
        "x-codegen-request-body-name" : "body"
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "LoginUser" : {
        "example" : {
          "password" : "password",
          "email" : "email"
        },
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "password" : {
            "format" : "password",
            "type" : "string"
          }
        },
        "required" : [ "email", "password" ],
        "type" : "object"
      },
      "LoginUserRequest" : {
        "example" : {
          "user" : {
            "password" : "password",
            "email" : "email"
          }
        },
        "properties" : {
          "user" : {
            "$ref" : "#/components/schemas/LoginUser"
          }
        },
        "required" : [ "user" ],
        "type" : "object"
      },
      "NewUser" : {
        "example" : {
          "password" : "password",
          "email" : "email",
          "username" : "username"
        },
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "password" : {
            "format" : "password",
            "type" : "string"
          }
        },
        "required" : [ "email", "password", "username" ],
        "type" : "object"
      },
      "NewUserRequest" : {
        "example" : {
          "user" : {
            "password" : "password",
            "email" : "email",
            "username" : "username"
          }
        },
        "properties" : {
          "user" : {
            "$ref" : "#/components/schemas/NewUser"
          }
        },
        "required" : [ "user" ],
        "type" : "object"
      },
      "User" : {
        "example" : {
          "image" : "image",
          "bio" : "bio",
          "email" : "email",
          "token" : "token",
          "username" : "username"
        },
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          }
        },
        "required" : [ "bio", "email", "image", "token", "username" ],
        "type" : "object"
      },
      "UserResponse" : {
        "example" : {
          "user" : {
            "image" : "image",
            "bio" : "bio",
            "email" : "email",
            "token" : "token",
            "username" : "username"
          }
        },
        "properties" : {
          "user" : {
            "$ref" : "#/components/schemas/User"
          }
        },
        "required" : [ "user" ],
        "type" : "object"
      },
      "UpdateUser" : {
        "example" : {
          "image" : "image",
          "bio" : "bio",
          "email" : "email",
          "token" : "token",
          "username" : "username"
        },
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UpdateUserRequest" : {
        "example" : {
          "user" : {
            "image" : "image",
            "bio" : "bio",
            "email" : "email",
            "token" : "token",
            "username" : "username"
          }
        },
        "properties" : {
          "user" : {
            "$ref" : "#/components/schemas/UpdateUser"
          }
        },
        "required" : [ "user" ],
        "type" : "object"
      },
...
      "GenericErrorModel" : {
        "properties" : {
          "errors" : {
            "$ref" : "#/components/schemas/GenericErrorModel_errors"
          }
        },
        "required" : [ "errors" ],
        "type" : "object"
      },
      "GenericErrorModel_errors" : {
        "properties" : {
          "body" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "body" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "Token" : {
        "description" : "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token xxxxxx.yyyyyyy.zzzzzz\n    \n",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}